requirements.txt
absl-py==0.11.0
astor==0.8.1
certifi==2020.6.20
cycler==0.10.0
gast==0.4.0
grpcio==1.33.2
h5py==2.10.0
importlib-metadata==2.0.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.1.0
Markdown==3.2.2
matplotlib==3.0.3
numpy==1.18.5
opencv-python==4.4.0.42
Pillow==7.2.0
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
scipy==1.4.1
six==1.15.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
Werkzeug==1.0.1
wincertstore==0.2
zipp==1.2.0
